+++ src/main/java/com/github/helf4ch/textstat/model/WordStat.java
@@ -3,7 +3,7 @@ package com.github.helf4ch.textstat.model;
 import org.springframework.data.annotation.Id;
 
 public class WordStat {
-  @Id private Long id;
+  @Id private Integer id;
   private String word;
   private Integer useCount;
 
@@ -12,11 +12,11 @@ public class WordStat {
     this.useCount = useCount;
   }
 
-  public Long getId() {
+  public Integer getId() {
     return id;
   }
 
-  public void setId(Long id) {
+  public void setId(Integer id) {
     this.id = id;
   }
