+++ src/main/java/com/github/helf4ch/textstat/repository/WordStatRepository.java
@@ -3,4 +3,4 @@ package com.github.helf4ch.textstat.repository;
 import com.github.helf4ch.textstat.model.WordStat;
 import org.springframework.data.repository.CrudRepository;
 
-public interface WordStatRepository extends CrudRepository<WordStat, Long> {}
+public interface WordStatRepository extends CrudRepository<WordStat, Integer> {}
